Make timeShiftQueryResponseOperator an util
